company orders screen broken up
