+++ code/client/economycraft/lib/main.dart
@@ -11,8 +11,8 @@ import 'package:supabase_flutter/supabase_flutter.dart';
 import 'package:economycraft/screens/holding/holdings_screen.dart';
 import 'package:economycraft/screens/player_overview/player_overview_screen.dart';
 import 'package:economycraft/screens/order/order_selection_screen.dart';
-import 'package:economycraft/screens/order/order_company_screen.dart';
-import 'package:economycraft/screens/order/order_user_screen.dart';
+import 'package:economycraft/screens/order/company/order_company_screen.dart';
+import 'package:economycraft/screens/order/user/order_user_screen.dart';
 import 'package:economycraft/screens/market_screen.dart';
 import 'package:economycraft/screens/player_profile_screen.dart';
 import 'package:economycraft/screens/company/company_page_screen.dart';
