+++ code/client/economycraft/lib/screens/order/company/widgets/company_order_details_dialog.dart
