+++ code/client/economycraft/lib/screens/order/user/widgets/order_details_dialog.dart
